% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correccion.R
\name{Correccion}
\alias{Correccion}
\title{Funcion tabla corregida}
\usage{
Correccion(TablaQ)
}
\arguments{
\item{TablaQ}{Tabla Quinquenal}
}
\value{
Tablacorregida
}
\description{
Funcion tabla corregida
}
