% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INU.R
\name{INU_f}
\alias{INU_f}
\title{Función que calcula el INU y devuelve su categoria}
\usage{
INU_f(TablaINU)
}
\arguments{
\item{TablaINU}{Tabla Quinquenal}
}
\value{
INU con su clasificación
}
\description{
Función que calcula el INU y devuelve su categoria
}
