% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quinquenios.R
\name{Quinquenios}
\alias{Quinquenios}
\title{Función para agrupar por grupos Quinquenales}
\usage{
Quinquenios(Censo)
}
\arguments{
\item{x}{Tabla M}
}
\value{
Tabla agrupada por grupos quinquenales
}
\description{
Función para agrupar por grupos Quinquenales
}
