% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INU.R
\name{INU}
\alias{INU}
\title{Función que calcula el INU}
\usage{
INU(Censo_Quinq)
}
\arguments{
\item{x}{Tabla Quinquenal}
}
\value{
Tabla con elementos para el calculo del INU así como el INU
}
\description{
Función que calcula el INU
}
